feat: TabMenuContext 에러 핸들링 추가 [BETA-113]
